requirements.txt
asgiref==3.7.1
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.1.3
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.2
urllib3==1.26.16
